fix(backend): Explicitly cast NumPy arrays to float64 in technical indicators
This change enforces the initialization of numpy arrays with `dtype=np.float64` in the `backend/utils/financial_math.py` calculations (`calculate_rsi`, `calculate_sma`, `calculate_ema`, `calculate_macd`, and `calculate_bbands`).

Previously, providing a list of pure integers to these functions could result in the creation of integer-type numpy arrays. Subsequent float divisions inside vectorized functions like `np.divide` would then crash with a `_UFuncOutputCastingError` because they could not cast float outputs back to int64. Explicitly creating the arrays as `np.float64` prevents this casting regression.

Also fixed some outdated unit tests in `tests/backend/test_financial_precision.py` and `tests/backend/test_quant_rebalance.py` that were improperly asserting `Decimal` returns instead of the float percentages outputted by the current `QuantEngine`.

In addition, fixed a `pytest.PytestRemovedIn9Warning` by replacing `@pytest.fixture` with `@pytest_asyncio.fixture` in `conftest.py`.

+++ backend/utils/financial_math.py
@@ -78,7 +78,7 @@ class TechnicalIndicators:
         """
         Calculates Relative Strength Index (RSI).
         """
-        data = np.array(prices)
+        data = np.array(prices, dtype=np.float64)
         if len(data) < period:
             return np.full(len(data), 50.0)
 
@@ -118,7 +118,7 @@ class TechnicalIndicators:
         """
         Calculates Simple Moving Average (SMA).
         """
-        arr = np.array(data)
+        arr = np.array(data, dtype=np.float64)
         if len(arr) < period:
             return np.zeros_like(arr)
 
@@ -155,7 +155,7 @@ class TechnicalIndicators:
         Calculates Exponential Moving Average (EMA).
         Uses SMA of first 'period' points for initialization to match Rust core.
         """
-        arr = np.array(data)
+        arr = np.array(data, dtype=np.float64)
         if len(arr) < period:
             return np.zeros_like(arr)
 
@@ -183,7 +183,7 @@ class TechnicalIndicators:
         """
         Calculates MACD Line, Signal Line, and Histogram.
         """
-        arr = np.array(data)
+        arr = np.array(data, dtype=np.float64)
 
         # 1. Rust Path
         if (backend == 'auto' or backend == 'rust') and RUST_CORE_AVAILABLE:
@@ -207,7 +207,7 @@ class TechnicalIndicators:
         """
         Calculates Bollinger Bands (Upper, Middle, Lower).
         """
-        arr = np.array(data)
+        arr = np.array(data, dtype=np.float64)
 
         # 1. MLX Path
         if (backend == 'auto' or backend == 'mlx') and MLX_AVAILABLE and len(arr) >= period:
